requirements.txt
Adafruit-Blinka==4.9.0
adafruit-circuitpython-busdevice==4.3.1
adafruit-circuitpython-tsl2561==3.3.1
Adafruit-PlatformDetect==2.11.1
Adafruit-PureIO==1.1.5
board==1.0
click==7.1.2
pyftdi==0.51.2
PyMySQL==0.9.3
pyserial==3.4
pyusb==1.0.2
PyYAML==5.3.1
rpi-ws281x==4.2.3
RPi.GPIO==0.7.0
sysv-ipc==1.0.1
w1thermsensor==1.3.0
